Allow any USB card to be accessed through DBus
If more than one sound card is available to ALSA, but only one is to be
used, we need to give DBus access to it.

Note: even with only 1 USB card, sometimes it even gets ALSA id 1
instead of 0. This happens if the system has previously booted with 2
cards (e.g. HifiBerry + USB) and on a subsequent boot the HifiBerry
device overlay is removed.

By granting access to cards >0, stuff should mostly keep working in
these varying scenarios.