Add browsepy service for mod-ui
This is required for the new File Manager feature of mod-ui. browsepy is a
moddevices fork which is patched for a specific directory layout. This change
only adds the 'Speaker Cabinets IRs' directory of that layout to support the
new cabsim plugin.

By setting MOD_USER_FILES_DIR to the browsepy root, the IR loader cabsim
Impulse Response dropdown will be populated with the contents of 'Speaker
Cabinets IRs'.